requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
statsmodels>=0.14.0
jupyter>=1.0.0
pytest>=7.3.0
black>=23.0.0
flake8>=6.0.0 
ucimlrepo>=0.0.3
xgboost==3.0.2
flask>=2.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
python-whois>=0.8.0
python-dotenv>=0.21.0
joblib>=1.1.0
python-dateutil>=2.8.2
tldextract>=5.1.1